pyproject.toml
[tool.poetry]
name = "tuya-python-examples"
version = "0.1.0"
description = ""
authors = ["danjamk <dan@risegardens.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tuya-connector-python = "^0.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
